+++ modules/transport-grpc/src/main/java/org/opensearch/transport/grpc/proto/request/search/suggest/TermSuggestionBuilderProtoUtils.java
@@ -29,7 +29,7 @@ public class TermSuggestionBuilderProtoUtils {
      * @return the corresponding TermSuggestionBuilder.SuggestMode
      * @throws IllegalArgumentException if the suggest_mode is invalid
      */
-    public static TermSuggestionBuilder.SuggestMode resolve(final SearchRequest.SuggestMode suggest_mode) {
+    public static TermSuggestionBuilder.SuggestMode resolve(final org.opensearch.protobufs.SuggestMode suggest_mode) {
         switch (suggest_mode) {
             case SUGGEST_MODE_ALWAYS:
                 return TermSuggestionBuilder.SuggestMode.ALWAYS;