+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased 3.x]
 ### Added
 - Upgrade opensearch-protobufs dependency to 0.7.0 and update transport-grpc module compatibility
+- Upgrade opensearch-protobufs dependency to 0.8.0 and update transport-grpc module compatibility
 
 ### Changed